Ajout Test demande de livraison
Ajout de ce test + fichier xml de test+ correctifs au code

+++ project/src/main/java/org/apache/project/modele/PlanDeVille.java
@@ -33,6 +33,6 @@ public class PlanDeVille extends Observable {
 	}
 	
 	public Intersection getIntersectionById(int id) {
-		return intersections.get(id);
+		return intersections.get((long)id);
 	}
 }
